Playing around with rust, figuring out the best way to do things.
This was an earlier project of mine that was in Python. And I thought
it would be a great way to learn Rust. I think I figured out a dry and
relatively simple way to do implement the rooms and necessary data
structures.

At some point, all the strings will have to be written regardless. For
now, it seems like a good idea to store that data in arrays or vectors.
I created one for each room struct property. To generate the room data,
we can loop over the length of any of those arrays and we can access
their elements via index. Seems like the least repetitive way to do it,
everything is written once, and accessed directly via reference.

The game data will be transferred in full, the current room data was
just testing out ideas.