Room Data has been completed. Data for all thirteen rooms has been added.
To me the implementation seems the most dry. Everything is declared once
and then reused throughout the program. A constants file was added
to the game data to handle static const data. Arrays were selected
as the data structure for storing the room names, items, and available
movements. Rather than repetitively hardcode the game data, the
`game_data::get_grooms` builds a json like struct <RoomData> containing
all the necessary individual room data objects or structs for the game.
Since we are looping, we can access all the other values for free if
they are stored in an indexable array.