Created a utils module to provide utility functions for receiving user
input and validating it. The input is guaranteed to be a non empty,
alphabetical string that can contain spaces. If the input is not valid,
the user will be prompted to enter a valid input.

The beginnings of a user input controller have also been created to
call the correct functions to handle the response. If the user provides
an input of go East, West, North, or South, then the movement
controller will be called. And so on.

But for now the main feature of retrieving user input has been created.