Finished the move between room feature.
A user can only move to a valid room. If the room doesn't have an
exit in the direction they want to move, they will be told that
the direction is invalid. And the next turn advances.

To implement this feature, a good majority of the game logic has been
implemented. A impl has been declared and I figured this was the
closest thing Rust has to a class.

The main method here is the start() function which loads a new game and
is the only publicly accessible function.The Game structure itself has
internal methods that are used to manipulate the game's state.

The controllers module was deleted and the logic is internal to the
Game structure. Data resides in the game_data module and extravagant
methods are declared in the utils module.