Added game ending and created the handler for the score.
A user wins the game when the have found all six items while avoiding
the enemy. When the game ends, the user is asked if they would like to
play again or if they want to quit. Some tweaks might need to be made
here and there, but the game is playable and have not encountered any
problems.